This commit adds support to duplicate bases layers
This is a an intermediate commit to duplicate the base layers for
different docker containers depending on which label(parent-cgroup)
they belong to. As of now, everything is supposed to be running from
"docker" cgroup, but this will certainly change in future to allow
specifying the cgroup from the command line.